Added support for parsing digital messages + adding basic parsing of I2C messages
But i had to change the way analog messages are read. The order of the command byte detection is important! Though it seems to work for the analog message parsing, it did not as streaming the same data through a digital message parser. When both are combined in the same loop - as within the c examples, p5, sharpduino or rhyduino - i tworks.